+++ release_version.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+VERSION="mbedtls"
+git checkout -b "$VERSION"
+git push --set-upstream origin "$VERSION"
+